+++ src/main/java/com/tristankechlo/toolleveling/utils/Names.java
@@ -9,6 +9,7 @@ public class Names {
 		public static final String UPGRADE_COST_MULTIPLIER = "upgradeCostMultiplier";
 		public static final String MIN_UPGRADE_COST = "minUpgradeCost";
 		public static final String ENCHANTMENT_BLACKLIST = "enchantmentBlacklist";
+		public static final String ENCHANTMENT_WHITELIST = "enchantmentWhitelist";
 		public static final String ENCHANTMENT_CAPS = "enchantmentCaps";
 		public static final String ITEM_VALUES = "itemValues";
 		public static final String ALLOW_WRONG_ENCHANTMENTS = "allowWrongEnchantments";
@@ -16,6 +17,7 @@ public class Names {
 		public static final String DEFAULT_ITEM_WORTH = "defaultItemWorth";
 		public static final String ALLOW_LEVELING_USELESS_ENCHANTMENTS = "allowLevelingOfUselessEnchantments";
 		public static final String ALLOW_LEVELING_BREAKING_ENCHANTMENTS = "allowLevelingOfBreakingEnchanments";
+		public static final String GLOBAL_ENCHANTMENT_CAP = "globalEnchantmentCap";
 	}
 
 }