+++ src/main/java/com/tristankechlo/toolleveling/utils/Utils.java
@@ -69,6 +69,9 @@ public class Utils {
 	}
 
 	public static boolean isEnchantmentAtCap(Enchantment enchantment, int level) {
+		if(ToolLevelingConfig.globalEnchantmentCap > 0) {
+			return level >= ToolLevelingConfig.globalEnchantmentCap;
+		}
 		if (ToolLevelingConfig.enchantmentCaps.containsKey(enchantment)) {
 			return (level >= ToolLevelingConfig.enchantmentCaps.get(enchantment));
 		}